requirements.txt
asgiref==3.3.1
boto3==1.17.0
botocore==1.20.0
certifi==2020.12.5
chardet==4.0.0
cloudinary==1.24.0
dj-database-url==0.5.0
Django==3.1.5
django-ckeditor==6.0.0
django-cloudinary-storage==0.3.0
django-crispy-forms==1.10.0
django-filter==2.4.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-storages==1.11.1
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
Pillow==8.1.0
psycopg2==2.8.6
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.15.0
python-magic==0.4.22
pytz==2020.5
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.3
whitenoise==5.2.0
